v1.4 Network request tests
closes #819
closes #820
